Implement compiling of PostgreSQL and benchmarking commits.
- Added function `generate_test`.
- Added fucntion `get_next_benchmark_run`.
- Added function `insert_result`.
- Added view `vresults`.
- Deleted function `new_run`.
- Deleted function `mark_run_as_started`.
- Changed `get_benchmark_test_ids` to use pseduo-random deterministic order.
- Change results.execution_time to float8 instead of numeric.
- Use timeit.f() instead of timeit.s() to get full precision.
  This is necessary since we want to measure stddev() andc confidence intervals.

+++ tables/results.sql
@@ -1,7 +1,7 @@
 CREATE TABLE catbench.results
 (
     id uuid NOT NULL DEFAULT pg_catalog.gen_random_uuid(),
-    execution_time numeric NOT NULL,
+    execution_time float8 NOT NULL,
     run_id uuid NOT NULL,
     test_id bigint NOT NULL,
     result_at timestamptz NOT NULL DEFAULT now(),
